requirements.txt
speechrecognition
pyttsx3
google-generativeai
datetime
webbrowser